+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## master (unreleased)
 
-Nothing here yet.
+- Small refactorings on the Gevena (Switzerland) holiday class.
 
 ## v7.1.1 (2019-11-22)